Fix #14225
