+++ core/templates/pages/learner-dashboard-page/learner-dashboard-icons.component.spec.ts
@@ -16,7 +16,7 @@
  * @fileoverview Unit tests for LearnerDashboardIconsComponent.
  */
 
-import { async, ComponentFixture, fakeAsync, TestBed, tick } from
+import { async, ComponentFixture, fakeAsync, flushMicrotasks, TestBed, tick } from
   '@angular/core/testing';
 
 import { LearnerDashboardIdsBackendApiService } from 'domain/learner_dashboard/learner-dashboard-ids-backend-api.service';
@@ -27,15 +27,23 @@ import { LearnerDashboardIconsComponent } from './learner-dashboard-icons.compon
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+import { NgbModal, NgbModalRef, NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { MockTranslatePipe } from 'tests/unit-test-utils';
 import { BrowserDynamicTestingModule } from '@angular/platform-browser-dynamic/testing';
 import { LearnerPlaylistModalComponent } from './modal-templates/learner-playlist-modal.component';
 
+class MockNgbModalRef {
+  componentInstance = {
+    activityId: null,
+    activityTitle: null,
+    activityType: null
+  };
+}
+
 describe('Learner Dashboard Icons Component', () => {
   let component: LearnerDashboardIconsComponent;
   let fixture: ComponentFixture<LearnerDashboardIconsComponent>;
-
+  let ngbModal: NgbModal;
   let learnerDashboardIdsBackendApiService:
     LearnerDashboardIdsBackendApiService;
   let learnerDashboardActivityBackendApiService:
@@ -67,6 +75,7 @@ describe('Learner Dashboard Icons Component', () => {
   }));
 
   beforeEach(() => {
+    ngbModal = TestBed.inject(NgbModal);
     fixture = TestBed.createComponent(LearnerDashboardIconsComponent);
     component = fixture.componentInstance;
     learnerDashboardIdsBackendApiService =
@@ -76,7 +85,7 @@ describe('Learner Dashboard Icons Component', () => {
     fixture.detectChanges();
     learnerPlaylistSpy = spyOn(
       learnerDashboardActivityBackendApiService,
-      'removeFromLearnerPlaylistModal'
+      'removeFromLearnerPlaylist'
     ).and.callThrough();
   });
 
@@ -728,6 +737,28 @@ describe('Learner Dashboard Icons Component', () => {
 
   it('should open a modal for removing exploration from' +
     ' learner playlist', fakeAsync(() => {
+    const learnerDashboardActivityIds = LearnerDashboardActivityIds
+      .createFromBackendDict({
+        incomplete_exploration_ids: [],
+        incomplete_collection_ids: [],
+        completed_exploration_ids: [],
+        completed_collection_ids: [],
+        exploration_playlist_ids: [],
+        collection_playlist_ids: [],
+        completed_story_ids: [],
+        learnt_topic_ids: [],
+        partially_learnt_topic_ids: [],
+        topic_ids_to_learn: [],
+        all_topic_ids: [],
+        untracked_topic_ids: []
+      });
+    component.learnerDashboardActivityIds = learnerDashboardActivityIds;
+    const modalSpy = spyOn(ngbModal, 'open').and.callFake((dlg, opt) => {
+      return (
+        { componentInstance: MockNgbModalRef,
+          result: Promise.resolve('url')
+        } as NgbModalRef);
+    });
     let activityId = '1';
     let activityTitle = 'Title';
     let activityType = 'exploration';
@@ -735,14 +766,37 @@ describe('Learner Dashboard Icons Component', () => {
     component.removeFromLearnerPlaylist(
       activityId, activityTitle, activityType);
     fixture.detectChanges();
-    tick(15000);
+    flushMicrotasks();
 
+    expect(modalSpy).toHaveBeenCalled();
     expect(learnerPlaylistSpy).toHaveBeenCalled();
   }
   ));
 
   it('should open a modal for removing collection from' +
     ' learner playlist', fakeAsync(() => {
+    const learnerDashboardActivityIds = LearnerDashboardActivityIds
+      .createFromBackendDict({
+        incomplete_exploration_ids: [],
+        incomplete_collection_ids: [],
+        completed_exploration_ids: [],
+        completed_collection_ids: [],
+        exploration_playlist_ids: [],
+        collection_playlist_ids: [],
+        completed_story_ids: [],
+        learnt_topic_ids: [],
+        partially_learnt_topic_ids: [],
+        topic_ids_to_learn: [],
+        all_topic_ids: [],
+        untracked_topic_ids: []
+      });
+    component.learnerDashboardActivityIds = learnerDashboardActivityIds;
+    const modalSpy = spyOn(ngbModal, 'open').and.callFake((dlg, opt) => {
+      return (
+        { componentInstance: MockNgbModalRef,
+          result: Promise.resolve('url')
+        } as NgbModalRef);
+    });
     let activityId = '1';
     let activityTitle = 'Title';
     let activityType = 'collection';
@@ -750,11 +804,63 @@ describe('Learner Dashboard Icons Component', () => {
     component.removeFromLearnerPlaylist(
       activityId, activityTitle, activityType);
     fixture.detectChanges();
-    tick(15000);
+    flushMicrotasks();
 
+    expect(modalSpy).toHaveBeenCalled();
     expect(learnerPlaylistSpy).toHaveBeenCalled();
   }
   ));
+
+  it(
+    'should open an ngbModal when removing from learner playlist' +
+    ' when calling removeFromLearnerPlaylistModal',
+    fakeAsync(() => {
+      const learnerDashboardActivityIds = LearnerDashboardActivityIds
+        .createFromBackendDict({
+          incomplete_exploration_ids: [],
+          incomplete_collection_ids: [],
+          completed_exploration_ids: [],
+          completed_collection_ids: [],
+          exploration_playlist_ids: [],
+          collection_playlist_ids: [],
+          completed_story_ids: [],
+          learnt_topic_ids: [],
+          partially_learnt_topic_ids: [],
+          topic_ids_to_learn: [],
+          all_topic_ids: [],
+          untracked_topic_ids: []
+        });
+      component.learnerDashboardActivityIds = learnerDashboardActivityIds;
+      const modalSpy = spyOn(ngbModal, 'open').and.callFake((dlg, opt) => {
+        return (
+        { componentInstance: MockNgbModalRef,
+          result: Promise.resolve('url')
+        } as NgbModalRef);
+      });
+      component.removeFromLearnerPlaylist(
+        '0', 'title', 'exploration');
+      expect(modalSpy).toHaveBeenCalled();
+      flushMicrotasks();
+      expect(learnerPlaylistSpy).toHaveBeenCalledWith(
+        '0', 'exploration', learnerDashboardActivityIds, 'url');
+    }));
+
+  it(
+    'should not remove anything from learner playlist when cancel ' +
+  'button is clicked when calling removeFromLearnerPlaylistModal',
+    fakeAsync(() => {
+      spyOn(ngbModal, 'open').and.callFake((dlg, opt) => {
+        return (
+        { componentInstance: MockNgbModalRef,
+          result: Promise.reject('success')
+        } as NgbModalRef);
+      });
+      component.removeFromLearnerPlaylist(
+        '0', 'title', 'collection');
+      flushMicrotasks();
+
+      expect(learnerPlaylistSpy).not.toHaveBeenCalled();
+    }));
 
   it('should add exploration to learner playlist', fakeAsync(() => {
     let activityId = '1';
