+++ core/templates/pages/learner-dashboard-page/learner-dashboard-icons.component.ts
@@ -26,6 +26,8 @@ import { LearnerDashboardActivityBackendApiService } from
   'domain/learner_dashboard/learner-dashboard-activity-backend-api.service';
 import { LearnerDashboardActivityIds } from
   'domain/learner_dashboard/learner-dashboard-activity-ids.model';
+import { LearnerPlaylistModalComponent } from './modal-templates/learner-playlist-modal.component';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
   selector: 'oppia-learner-dashboard-icons',
@@ -55,7 +57,8 @@ export class LearnerDashboardIconsComponent implements OnInit {
     private learnerDashboardIdsBackendApiService:
       LearnerDashboardIdsBackendApiService,
     private learnerDashboardActivityBackendApiService:
-      LearnerDashboardActivityBackendApiService
+      LearnerDashboardActivityBackendApiService,
+    private ngbModal: NgbModal,
   ) {}
 
   ngOnInit(): void {
@@ -171,12 +174,37 @@ export class LearnerDashboardIconsComponent implements OnInit {
     }
   }
 
+  // This function will open a modal to remove an exploration
+  // from the 'Play Later' list in the Library Page.
   removeFromLearnerPlaylist(
       activityId: string, activityTitle: string, activityType: string): void {
-    this.learnerDashboardActivityBackendApiService
-      .removeFromLearnerPlaylistModal(
-        activityId, activityTitle, activityType,
-        this.learnerDashboardActivityIds);
+    // This following logic of showing a modal for confirmation previously
+    // resided in learnerDashboardActivityBackendApiService. However, in
+    // issue #14225, we noticed some errors with dynamic component creation.
+    // The componentFactoryResolver in the service was from AppModule and not
+    // the page specific module. This makes sense as the we provide all
+    // services in root (As specified by the providedIn: 'root'). So the
+    // injector used is the root injector and not the page module injector.
+    // The entry components specified in page module won't be available when
+    // we use the root injector.
+    // TODO(14290): Find a better way to refactor code that opens modals
+    // into new services that use the page specific injector rather than
+    // the root injector.
+    const modelRef = this.ngbModal.open(
+      LearnerPlaylistModalComponent, {backdrop: true});
+    modelRef.componentInstance.activityId = activityId;
+    modelRef.componentInstance.activityTitle = activityTitle;
+    modelRef.componentInstance.activityType = activityType;
+    modelRef.result.then((playlistUrl) => {
+      this.learnerDashboardActivityBackendApiService
+        .removeFromLearnerPlaylist(
+          activityId, activityType,
+          this.learnerDashboardActivityIds, playlistUrl);
+    }, () => {
+      // Note to developers:
+      // This callback is triggered when the Cancel button is clicked.
+      // No further action is needed.
+    });
   }
 }
